Update!
I'm going to start migrating away from NixOS and back to MacOS/Arch
Linux or maybe Void. The complexity of Nix is not something I want to
have in my daily life. If Nix makes it easy to build a certain project
then cool but my OS is not something I switch so often that I need to
automate everything at the cost of layers and layers of abstractions